+++ src/api/subcategory.ts
@@ -1,42 +1,55 @@
-import type { ItemType, SubcategoriesType } from '@app/types/list.types';
+'use server';
 
-export async function getSubcategoryItems(
-  id: string
-): Promise<{ subcategoryItems: ItemType[] } | undefined> {
-  try {
-    const res = await fetch(
-      `${process.env.NEXT_PUBLIC_API_URL}subcategory-items/subcategory/${id}`,
-      {
-        method: 'GET',
-        headers: {
-          Authorization: `Bearer ${process.env.NEXT_PUBLIC_TOKEN}`,
-          'Content-Type': 'application/json',
-        },
-        cache: 'no-store',
-      }
-    );
-    return res.json();
-  } catch (e) {
-    console.log('error', e);
-  }
-}
+import { revalidateTag } from 'next/cache';
+import { cookies } from 'next/headers';
+
+import type { SubcategoriesType } from '@app/types/list.types';
 
 export async function getAllSubcategories(): Promise<
   { subcategories: SubcategoriesType[] } | undefined
 > {
   try {
-    const res = await fetch(
-      `${process.env.NEXT_PUBLIC_API_URL}all-subcategories`,
-      {
-        method: 'GET',
-        headers: {
-          Authorization: `Bearer ${process.env.NEXT_PUBLIC_TOKEN}`,
-          'Content-Type': 'application/json',
-        },
-        cache: 'no-store',
+    const token = cookies().get('token');
+    if (token && token.value) {
+      if (token) {
+        const res = await fetch(
+          `${process.env.NEXT_PUBLIC_API_URL}all-subcategories`,
+          {
+            method: 'GET',
+            headers: {
+              Authorization: `Bearer ${process.env.NEXT_PUBLIC_TOKEN}`,
+              'Content-Type': 'application/json',
+            },
+            next: { tags: ['subcategory'] },
+          }
+        );
+        return await res.json();
       }
-    );
-    return res.json();
+    }
+  } catch (e) {
+    console.log('error', e);
+  }
+}
+
+export async function getSubcategory(
+  id: string
+): Promise<{ subcategory: SubcategoriesType } | undefined> {
+  try {
+    const token = cookies().get('token');
+    if (token && token.value) {
+      const res = await fetch(
+        `${process.env.NEXT_PUBLIC_API_URL}subcategories/${id}`,
+        {
+          method: 'GET',
+          headers: {
+            Authorization: `Bearer ${process.env.NEXT_PUBLIC_TOKEN}`,
+            'Content-Type': 'application/json',
+          },
+          next: { tags: ['subcategory'] },
+        }
+      );
+      return await res.json();
+    }
   } catch (e) {
     console.log('error', e);
   }
@@ -47,18 +60,23 @@ export async function createSubcategory(
   data: SubcategoriesType
 ): Promise<{ subcategoryItem: SubcategoriesType } | undefined> {
   try {
-    const res = await fetch(
-      `${process.env.NEXT_PUBLIC_API_URL}subcategories/${categoryId}`,
-      {
-        method: 'POST',
-        headers: {
-          Authorization: `Bearer ${process.env.NEXT_PUBLIC_TOKEN}`,
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify(data),
-      }
-    );
-    return res.json();
+    const token = cookies().get('token');
+    if (token && token.value) {
+      const res = await fetch(
+        `${process.env.NEXT_PUBLIC_API_URL}subcategories/${categoryId}`,
+        {
+          method: 'POST',
+          headers: {
+            Authorization: `Bearer ${process.env.NEXT_PUBLIC_TOKEN}`,
+            'Content-Type': 'application/json',
+          },
+          body: JSON.stringify(data),
+        }
+      );
+      const result = await res.json();
+      revalidateTag('subcategory');
+      return result;
+    }
   } catch (e) {
     console.log('error', e);
   }
@@ -69,18 +87,23 @@ export async function updateSubcategory(
   data: SubcategoriesType
 ): Promise<{ subcategoryItem: SubcategoriesType } | undefined> {
   try {
-    const res = await fetch(
-      `${process.env.NEXT_PUBLIC_API_URL}subcategories/${categoryId}`,
-      {
-        method: 'PUT',
-        headers: {
-          Authorization: `Bearer ${process.env.NEXT_PUBLIC_TOKEN}`,
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify(data),
-      }
-    );
-    return res.json();
+    const token = cookies().get('token');
+    if (token && token.value) {
+      const res = await fetch(
+        `${process.env.NEXT_PUBLIC_API_URL}subcategories/${categoryId}`,
+        {
+          method: 'PUT',
+          headers: {
+            Authorization: `Bearer ${process.env.NEXT_PUBLIC_TOKEN}`,
+            'Content-Type': 'application/json',
+          },
+          body: JSON.stringify(data),
+        }
+      );
+      const result = await res.json();
+      revalidateTag('subcategory');
+      return result;
+    }
   } catch (e) {
     console.log('error', e);
   }
@@ -88,34 +111,17 @@ export async function updateSubcategory(
 
 export async function deleteSubcategory(id: string): Promise<void> {
   try {
-    await fetch(`${process.env.NEXT_PUBLIC_API_URL}subcategories/${id}`, {
-      method: 'DELETE',
-      headers: {
-        Authorization: `Bearer ${process.env.NEXT_PUBLIC_TOKEN}`,
-        'Content-Type': 'application/json',
-      },
-    });
-  } catch (e) {
-    console.log('error', e);
-  }
-}
-
-export async function getSubcategory(
-  id: string
-): Promise<{ subcategory: SubcategoriesType } | undefined> {
-  try {
-    const res = await fetch(
-      `${process.env.NEXT_PUBLIC_API_URL}subcategories/${id}`,
-      {
-        method: 'GET',
+    const token = cookies().get('token');
+    if (token && token.value) {
+      await fetch(`${process.env.NEXT_PUBLIC_API_URL}subcategories/${id}`, {
+        method: 'DELETE',
         headers: {
           Authorization: `Bearer ${process.env.NEXT_PUBLIC_TOKEN}`,
           'Content-Type': 'application/json',
         },
-        cache: 'no-store',
-      }
-    );
-    return res.json();
+      });
+      revalidateTag('subcategory');
+    }
   } catch (e) {
     console.log('error', e);
   }
